Support QML and Web content in overlay windows
