+++ CMakeGraphvizOptions.cmake
@@ -1,0 +1,1 @@
+set(GRAPHVIZ_EXTERNAL_LIBS FALSE)
