+++ interface/resources/qml/QmlWebWindow.qml
@@ -1,9 +1,6 @@
-
 import QtQuick 2.3
 import QtQuick.Controls 1.2
 import QtWebEngine 1.1
-import QtWebChannel 1.0
-import QtWebSockets 1.0
 
 import "controls"
 import "styles"
@@ -13,6 +10,7 @@ VrDialog {
     HifiConstants { id: hifi }
     title: "WebWindow"
     resizable: true
+	enabled: false
     // Don't destroy on close... otherwise the JS/C++ will have a dangling pointer
     destroyOnCloseButton: false
     contentImplicitWidth: clientArea.implicitWidth
@@ -24,18 +22,18 @@ VrDialog {
     function stop() {
         webview.stop();
     }
-    
 
     Component.onCompleted: {
-        enabled = true
-        console.log("Web Window Created " + root);
+        // Ensure the JS from the web-engine makes it to our logging
         webview.javaScriptConsoleMessage.connect(function(level, message, lineNumber, sourceID) {
             console.log("Web Window JS message: " + sourceID + " " + lineNumber + " " +  message);
         });
+
+	    // Required to support clicking on "hifi://" links
         webview.loadingChanged.connect(handleWebviewLoading) 
     }
 
-
+    // Required to support clicking on "hifi://" links
     function handleWebviewLoading(loadRequest) {
         if (WebEngineView.LoadStartedStatus == loadRequest.status) {
             var newUrl = loadRequest.url.toString();
@@ -56,6 +54,7 @@ VrDialog {
             id: webview
             url: root.source
             anchors.fill: parent
+
             onUrlChanged: {
                 var currentUrl = url.toString();
                 var newUrl = urlFixer.fixupUrl(currentUrl);
@@ -63,6 +62,7 @@ VrDialog {
                     url = newUrl;
                 }
             }
+
             profile: WebEngineProfile {
                 id: webviewProfile
                 httpUserAgent: "Mozilla/5.0 (HighFidelityInterface)"
